# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
  github_com_Daty26_todo-app_internal_core_transport_http_types.Nullable-bool:
    properties:
      set:
        type: boolean
      value:
        type: boolean
    type: object
  github_com_Daty26_todo-app_internal_core_transport_http_types.Nullable-string:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
  internal_features_statistics_transport_http.GetStatisticsResponse:
    properties:
      task_completed:
        example: 10
        type: integer
      task_completed_rate:
        example: 20
        type: number
      task_created:
        example: 50
        type: integer
      tasks_average_completion_time:
        example: 1m30s
        type: string
    type: object
  internal_features_tasks_transport_http.CreateTaskRequest:
    properties:
      author_user_id:
        type: integer
      description:
        maxLength: 1000
        minLength: 1
        type: string
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - author_user_id
    - title
    type: object
  internal_features_tasks_transport_http.CreateTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.GetTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.PatchTaskRequest:
    properties:
      completed:
        $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_types.Nullable-bool'
      description:
        $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_types.Nullable-string'
      title:
        $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_types.Nullable-string'
    type: object
  internal_features_tasks_transport_http.PatchTaskResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_tasks_transport_http.TaskDTOResponse:
    properties:
      author_user_id:
        type: integer
      completed:
        type: boolean
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      version:
        type: integer
    type: object
  internal_features_users_transport_http.CreateUserRequest:
    properties:
      full_name:
        example: Max Verstappen
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+76557078867"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transport_http.CreateUserResponse:
    properties:
      full_name:
        example: Max Verstappen
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+76557078867"
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.GetUserResponse:
    properties:
      full_name:
        example: Max Verstappen
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+76557078867"
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.PatchUserRequest:
    properties:
      full_name:
        example: Max Verstappen
        type: string
      phone_number:
        example: "+770765343112"
        type: string
    type: object
  internal_features_users_transport_http.PatchUserResponse:
    properties:
      full_name:
        example: Max Verstappen
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+76557078867"
        type: string
      version:
        example: 1
        type: integer
    type: object
  internal_features_users_transport_http.UserDTOResponse:
    properties:
      full_name:
        example: Max Verstappen
        type: string
      id:
        example: 1
        type: integer
      phone_number:
        example: "+76557078867"
        type: string
      version:
        example: 1
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Todo Application RES-API scheme
  title: Todo API
  version: "1.0"
paths:
  /statistics:
    get:
      description: Get task statistics with optional user and date filters
      parameters:
      - description: ID of the user to filter statistics by
        in: query
        name: user_id
        type: integer
      - description: Start date in YYYY-MM-DD format
        in: query
        name: from
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully received statistics
          schema:
            $ref: '#/definitions/internal_features_statistics_transport_http.GetStatisticsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get statistics
      tags:
      - statistics
  /tasks:
    get:
      description: View the list of tasks with optional user filter and pagination
      parameters:
      - description: ID of the user that owns the tasks
        in: query
        name: user_id
        type: integer
      - description: The size of the page with tasks
        in: query
        name: limit
        type: integer
      - description: Offset of the page with tasks
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully received the list of tasks
          schema:
            items:
              $ref: '#/definitions/internal_features_tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create new task in the system
      parameters:
      - description: CreateTask request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.CreateTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete an existing task by id
      parameters:
      - description: ID of the task that needs to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted task
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get specific task by id
      parameters:
      - description: ID of the task
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task was successfully found
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.GetTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get task
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Partially update an existing task by id
      parameters:
      - description: ID of the task that needs to be patched
        in: path
        name: id
        required: true
        type: integer
      - description: PatchTask request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully patched task
          schema:
            $ref: '#/definitions/internal_features_tasks_transport_http.PatchTaskResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Patch task
      tags:
      - tasks
  /users:
    get:
      description: View the list of users with optional pagination
      parameters:
      - description: The size of the page with users
        in: query
        name: limit
        type: integer
      - description: Offset of the page with users
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully received the list of users
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transport_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: List of users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user in the system
      parameters:
      - description: CreateUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletion of already existing user by id
      parameters:
      - description: ID of the user that needs to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted user
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not Found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: User deletion
      tags:
      - users
    get:
      description: Get specific user by id
      parameters:
      - description: ID of the user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User was successfully found
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Get user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Partially update an existing user by id
      parameters:
      - description: ID of the user that needs to be patched
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully patched user
          schema:
            $ref: '#/definitions/internal_features_users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_Daty26_todo-app_internal_core_transport_http_response.ErrorResponse'
      summary: Patch user
      tags:
      - users
swagger: "2.0"